Fix 11 QA findings from FINAL_ALT end-to-end re-verification
Resolve all in-scope runtime defects identified by QA, verified end-to-end
against live services (bank-core:8080, Customer Services:8081, Payment:8082,
PostgreSQL cbsa). Frozen z/OS Connect wire contract fully preserved
(zero @JsonProperty / Jackson envelope changes). BigDecimal scale-2 money and
fixed-width String identifiers preserved through all layers.

bank-core API / validation / error handling:
- Issue 6 (MAJOR): GlobalExceptionHandler now maps DataIntegrityViolationException
  to a sanitized HTTP 400 (WARN log, no SQL/stack leak) so boundary money values
  (e.g. NUMERIC(12,2) overflow on PUT /makepayment/dbcr) no longer return HTTP 500.
- Issue 7 (MAJOR): @NotNull on UpdaccJson commAccno/commAccountType/commInterestRate/
  commOverdraft so CREACC-style alias field names fail with clean HTTP 400 validation
  instead of reaching a JPA NOT NULL constraint (HTTP 500).
- Issue 8 (MINOR): @ExceptionHandler(HttpMediaTypeNotAcceptableException) returns a
  clean 406 (no body) and avoids the recursive @ExceptionHandler failure on
  Accept: text/plain.
- Issue 9 (INFO): Javadoc on CrecustJson/CreateCustomerJson/CustomerService documents
  that the title is parsed from the first token of CommName and that CommTitle is not
  a frozen-contract field (no behavior change; frozen-contract leniency preserved).

Security / dependency posture:
- Issue 10 (MAJOR): bump spring-boot-starter-parent 3.5.11 -> 3.5.14 across bank-core
  and both interface modules to remediate High CVE-2026-40973 (GHSA-wwpq-f5c3-7hvx);
  CVE pin reconciliation (keep postgresql/tomcat/logback, drop pins the 3.5.14 BOM
  now manages at >= fixed versions).
- Issue 11 (MINOR): revert lodash 4.17.23 -> 4.17.21 in frontend package.json +
  yarn.lock so the lockfile is consistent and dependency audit can run.

Customer Services / Payment interface preserved-consumer fidelity:
- Issue 3 (MAJOR): robust review-date/DOB parsing in CS UpdcustJson (parseDateToInt:
  strip non-digits, default 0, never throw) so a populated review date no longer
  triggers NumberFormatException / HTTP 500 on /updatecust.
- Issue 4 (MAJOR): CS customer-number identifier fields int -> String so valid
  10-digit IDs near the upper bound no longer overflow / InvalidFormatException.
- Issue 5 (MINOR): interface DTO money float/int -> BigDecimal and customer-number
  identifiers int -> String across CS + Payment DTOs; debit arithmetic uses
  BigDecimal.negate() (no double). Account numbers (8) and sort codes (6) stay int.

Self-introduced regression (found and fixed during runtime re-verification):
- The Issue 10 Spring Boot bump moved BOM-managed thymeleaf-spring6 3.1.3 -> 3.1.5
  while the Payment pom pinned thymeleaf core at 3.1.2, causing
  NoSuchMethodError (StandardExpressionExecutionContext.getRestrictExternalAccess)
  on every rendered Thymeleaf page. Removed the explicit thymeleaf core pin and the
  hardcoded spring-boot-starter-thymeleaf version so both artifacts are BOM-managed
  in lockstep at 3.1.5 (matching Customer Services). Verified root GET / and /submit
  render/respond HTTP 200 with zero stack traces.

Documentation:
- Issue 14 (MINOR): repoint README Liberty UI user-guide link to the existing
  carbonReactUI guide.
- Issue 15 (MINOR): repoint architecture guide image to the existing CarbonReactUI
  architecture diagram.

Validation: ./mvnw -B clean package BUILD SUCCESS (4/4 modules); bank-core
164 tests, 0 failures/errors/skips. Deferred (explicit AAP exclusions, not fixed):
Issues 1, 2, 12, 13 (webui WAR deployability + React redesign/responsive/a11y) per
AAP sec 0.1.1 / 0.2.2 / 0.3.4.

+++ src/Z-OS-Connect-Customer-Services-Interface/src/main/java/com/ibm/cics/cip/bank/springboot/customerservices/jsonclasses/createaccount/CreateAccountForm.java
@@ -3,6 +3,8 @@
 /*                                                                        */
 package com.ibm.cics.cip.bank.springboot.customerservices.jsonclasses.createaccount;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.*;
 
 public class CreateAccountForm
@@ -18,7 +20,10 @@ public class CreateAccountForm
 
 	private int overdraftLimit;
 
-	private float interestRate;
+	// QA Issue 5: interest rate is monetary precision; AAP requires BigDecimal
+	// (scale-2) and prohibits float/double. Spring MVC binds the form value
+	// straight into BigDecimal.
+	private BigDecimal interestRate;
 
 
 	public int getOverdraftLimit()
@@ -33,13 +38,13 @@ public class CreateAccountForm
 	}
 
 
-	public float getInterestRate()
+	public BigDecimal getInterestRate()
 	{
 		return interestRate;
 	}
 
 
-	public void setInterestRate(float interestRate)
+	public void setInterestRate(BigDecimal interestRate)
 	{
 		this.interestRate = interestRate;
 	}
